+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
 import java.time.LocalDate;
-import roomescape.domain.reservation.Reservation;
+import roomescape.domain.Reservation;
 
 public record ReservationRequest(String name, LocalDate date, Long timeId) {
     public Reservation toReservation() {
